feat(InputView): InputView 인터페이스화
테스트에서 유용성 및 확장성 고려
