+++ src/main/java/lotto/view/InputView.java
@@ -1,39 +1,10 @@
 package lotto.view;
 
-import lotto.constants.string.InputMessage;
+public interface InputView {
 
-import static camp.nextstep.edu.missionutils.Console.readLine;
+    String getInputPrice();
 
-public class InputView {
+    String getWinningComponent();
 
-    private static InputView inputView;
-
-    private InputView() {
-    }
-
-    public static InputView getInstance() {
-        if (inputView == null) {
-            inputView = new InputView();
-        }
-        return inputView;
-    }
-
-    public String getInputPrice() {
-        System.out.println(InputMessage.PRICE.getInstance());
-        return readLine();
-    }
-
-    public String getWinningComponent() {
-        System.out.println();
-        System.out.println(InputMessage.WINNING_NUMBER.getInstance());
-        return readLine();
-    }
-
-
-    public String getBonusComponent() {
-        System.out.println();
-        System.out.println(InputMessage.BONUS_NUMBER.getInstance());
-        return readLine();
-    }
-
+    String getBonusComponent();
 }
